fix(#715): move rpc_transport.py from core/ to services/ — concrete HTTP transport not kernel
NexusRPCTransport is a concrete HTTP-based RPC transport implementation, not a kernel
ABC. Per KERNEL-ARCHITECTURE.md, concrete transport implementations are drivers/services,
not kernel primitives. Moved to nexus.services.rpc_transport and updated test import.
